\name{team10stats799-package}
\alias{team10stats799-package}
\alias{team10stats799}
\docType{package}
\title{
\packageTitle{team10stats799}
}
\description{
\packageDescription{team10stats799}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{team10stats799}
\packageIndices{team10stats799}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{team10stats799}

Maintainer: \packageMaintainer{team10stats799}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
